LG-2041: renamed InterstitialController to Users::AuthorizationConfirmationController
