+++ config/locales/user_authorization_confirmation/fr.yml
@@ -1,6 +1,6 @@
 ---
-en:
-  interstitial:
+fr:
+  user_authorization_confirmation:
     currently_logged_in: You are currently logged in as
     not_you: Not you?
     continue: Continue
